refactor: 메뉴 그룹 도메인 이동
