+++ src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.menugroup.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupRequest {
     private String name;
